Move Provider classes to providers package
